make nadm file and creat wve_time in fin_diff.cc

+++ c++/arrayman.h
@@ -1,14 +1,14 @@
 #ifndef ARRAYMAN_H
 #define ARRAYMAN_H
-int len(float *a);
-float *mat_to_vek(float **arr,const int nrows, const int ncols);
-float **alloc_mat(const int nrows, const int ncols);
-void free_mat_mem(float **mat);
-void print_mat(float **mat,const int nrows,const int ncols);
-void print_vek(float *mat,const int nelmts);
-void w_file_mat(char *filename, float **P,int nt,int nrows, int ncols);
-void w_vtk(char *filename,float *Vel_mod, float **P,float h,int nt,int nx, int ny, int nz);
-void w_csv(char *filename,float *Vel_mod, float **P,float dt,float h,int nt,int nx, int ny, int nz);
-void w_dat(char *filename,float *Vel_mod, float **P,float dt,float h,int nt,int nx, int ny, int nz);
+int len(double *a);
+double *mat_to_vek(double **arr,const int nrows, const int ncols);
+double **alloc_mat(const int nrows, const int ncols);
+void free_mat_mem(double **mat);
+void print_mat(double **mat,const int nrows,const int ncols);
+void print_vek(double *mat,const int nelmts);
+void w_file_mat(char *filename, double **P,int nt,int nrows, int ncols);
+void w_vtk(char *filename,double *Vel_mod, double **P,double h,int nt,int nx, int ny, int nz);
+void w_csv(char *filename,double *Vel_mod, double **P,double dt,double h,int nt,int nx, int ny, int nz);
+void w_dat(char *filename,double *Vel_mod, double **P,double dt,double h,int nt,int nx, int ny, int nz);
 #endif
