+++ c++/arrayopr.cc
@@ -1,26 +1,26 @@
-float *vek_addition(float *A, float *B,const int nelmts){
-	float *C=new float[nelmts];
+double *vek_addition(double *A, double *B,const int nelmts){
+	double *C=new double[nelmts];
 	for (int i=0;i<nelmts;i++){
 			C[i]=A[i]+B[i];
 	}
 	return C;
 }
-float *vek_subtraction(float *A, float *B ,const int nelmts){
-	float *C=new float[nelmts];
+double *vek_subtraction(double *A, double *B ,const int nelmts){
+	double *C=new double[nelmts];
 	for (int i=0;i<nelmts;i++){
 			C[i]=A[i]-B[i];
 	}
 	return C;
 }
-float *vek_elmwise_mult(float *A, float *B,const int nelmts){
-	float *C=new float[nelmts];
+double *vek_elmwise_mult(double *A, double *B,const int nelmts){
+	double *C=new double[nelmts];
 	for (int i=0;i<nelmts;i++){
 			C[i]=A[i]*B[i];
 	}
 	return C;
 }
-float *scalar_mult(const float scalar_const,float *A,const int nelmts){
-	float *C=new float[nelmts];
+double *scalar_mult(const double scalar_const,double *A,const int nelmts){
+	double *C=new double[nelmts];
 	for (int i=0;i<nelmts;i++){
 			C[i]=scalar_const*A[i];
 	}
