+++ c++/arrayopr.h
@@ -1,7 +1,7 @@
 #ifndef ARRAYOPR_H
 #define ARRAYOPR_H
-float *vek_addition(float *A, float *B,const int nelmts);
-float *vek_subtraction(float *A, float *B,const int nelmts);
-float *vek_elmwise_mult(float *A, float *B,const int nelmts);
-float *scalar_mult(float scalar_const,float *A,const int nelmts);
+double *vek_addition(double *A, double *B,const int nelmts);
+double *vek_subtraction(double *A, double *B,const int nelmts);
+double *vek_elmwise_mult(double *A, double *B,const int nelmts);
+double *scalar_mult(double scalar_const,double *A,const int nelmts);
 #endif
