+++ c++/source_function.cc
@@ -1,7 +1,7 @@
 #include <math.h>
-float ricker_wavelet(float freq, float t){
-	float f;
-	float pi=22.0/7.0;
+double ricker_wavelet(double freq, double t){
+	double f;
+	double pi=22.0/7.0;
     f = sin(2.0 * pi * freq * t) *exp(-4.0* (pi*freq*t)*(pi*freq*t)/ 16.0);
     return f;
 }
